fix(identity): application 모듈 Spring 배선 분리 #48
