+++ systems/identity/identity-application/deps.bzl
@@ -1,5 +1,4 @@
 KOTLIN_DEPS = [
-    "@maven//:org_springframework_boot_spring_boot_starter",
     "//systems/identity/identity-domain:main",
     "@maven//:io_jsonwebtoken_jjwt_api",
     "@maven//:io_jsonwebtoken_jjwt_impl",
