+++ systems/identity/identity-bootstrap/deps.bzl
@@ -5,6 +5,7 @@ SPRING_DEPS = [
 ]
 
 KOTLIN_DEPS = [
+    "@maven//:org_springframework_boot_spring_boot_autoconfigure",
     "@maven//:org_jetbrains_kotlin_kotlin_reflect",
     "@maven//:com_fasterxml_jackson_module_jackson_module_kotlin",
     "@maven//:io_jsonwebtoken_jjwt_api",
@@ -15,6 +16,7 @@ KOTLIN_DEPS = [
 TEST_DEPS = [
     "@maven//:org_springframework_boot_spring_boot_starter_test",
     "@maven//:junit_junit",
+    "@maven//:org_testcontainers_testcontainers",
 ]
 
 MODULE_DEPS = SPRING_DEPS + KOTLIN_DEPS
